Break out partial functions into their own module
